LIRI-Bot
App Description

LIRI is like iPhone's SIRI. However, while SIRI is a Speech
Interpretation and Recognition Interface, LIRI is a Language
Interpretation and Recognition Interface. LIRI will be a command line
node app that takes in parameters and gives you back data.

Description on how to use the app

Once cd-ing into the LIRI-Bot directory, type in: node liri.js
INSERTCOMMAND
List of commands:
my-tweets
spotify-this-song
movie-this
do-what-it-says
If no song is typed in after the spotify-this-song, a default song will
be searched.
If no movie is typed in after the movie-this, a default movie will be
searched.